Improve animation system and fix avatar flickering
The animation system now distinguishes between
animationDuration and animationRelease, where:
- animationDuration: the real duration of the animation.
The callback function, if passed, will be executed after this
time.
- animationRelease: the time after which the animation is
released (animation class removed, original position
restored). By default, it's the same as animationDuration,
but the animation can be forced to be released some
time after it's really finished.

The use of this new functionality fixes avatar flickering
due to a mismatch between slideOut and moveToCurrent
animation durations.